refactor(edit-model): extract palette width resizing handling into hook
