+++ src/EditModel/EditDataEntryForm/clampPaletteWidth.js
@@ -1,0 +1,1 @@
+export const clampPaletteWidth = width => Math.min(750, Math.max(width, 250))
